Newest version
Nearly the final version. Has exception handling as well as logging built in. Must have a file called logs.txt in the same folder as the python running. Must also have an excel file called empty_test.xlsx, the name of which can be changed as long as it is also changed in the code. Will break if given the recaptcha that identifies you as a bot. In the logs will be the page name.